backend/requirements.txt
fastapi
uvicorn
google-generative-ai
python-dotenv
pydantic
pydantic-settings
cors
